+++ packages/atomic/src/components/search/lazy-index.ts
@@ -3,8 +3,14 @@ export default {
   'atomic-pager': async () => await import('./atomic-pager/atomic-pager.js'),
   'atomic-query-summary': async () =>
     await import('./atomic-query-summary/atomic-query-summary.js'),
+  'atomic-relevance-inspector': async () =>
+    await import('./atomic-relevance-inspector/atomic-relevance-inspector.js'),
   'atomic-results-per-page': async () =>
     await import('./atomic-results-per-page/atomic-results-per-page.js'),
+  'atomic-search-box-recent-queries': async () =>
+    await import(
+      './atomic-search-box-recent-queries/atomic-search-box-recent-queries.js'
+    ),
   'atomic-search-interface': async () =>
     await import('./atomic-search-interface/atomic-search-interface.js'),
   'atomic-search-layout': async () =>